WIP: checkpointing normalization and arithmetic add code
