+++ tests/bfloat/api/constexpr.cpp
@@ -90,7 +90,7 @@ try {
 	
 	using Real = bfloat<12, 2>;
 	Real a;
-	a.debug();
+	a.constexprParameters();
 
 	TestConstexprConstruction<Real>();
 	TestConstexprAssignment<Real>();
